requirements.txt
fastapi
uvicorn[standard]
httpx
twikit
python-dotenv
